Revamp error handling and port tests to Jest
Currently the internal error handling sends out errors as plain
text which is not usually desirable especially if the client is
expecting JSON. Send errors as JSON if the client accepts any
content type as a response or requests for a JSON response body.

Closes #4

Remove magic error handling when a `statusCode` is present in an
`Error` object as that seems to cause more trouble than good in
general, e.g. accidentally leaking backend API errors and secrets.
As it might be sometimes desirable to short-circuit and throw a
non-500 error response let's implement a new `MicriError` object
that supports exactly that but with a well defined API that will
never leak any secrets.

Closes #1