Split the index.ts and create single entry point to Micri
The separate executable file is unnecessary as we can just easily
combine both the executable and a the library main in the single
file. This will enable us to minify and pack `Micri` into a
single file once we find a feasible solution for that.

Splitting the code improves readability and maintanability.

Also remove undocumented deprecated args.