Use handler-agent for better with-worker performance
Currently `with-worker` causes `worker-wrapper` to create temp
HTTP server which is used to serve the request over TCP. This
naturally causes some unnecessary latency as the data must go thru
the kernel in order to get it back to the handler. While
`handler-agent` still creates a temporary server it uses a clever
way to pipe the request directly to the handler thru Node.js's
internals, avoiding all the syscalls and random latency penalties.